requirements.txt
Flask
pymongo
bcrypt
python-dotenv
requests
google-generativeai
